Fix create2Workaround tests
